Housekeeping of the various admin scripts.

+++ complete_reset.bash
@@ -1,8 +1,18 @@
 #!/bin/bash
 
+# A script to OBLITERATE and repopulate an EXISTING database, IFF the 
+# database is SQLite.
+
 # Make sure the script gets run in the same directory as the script itself.
 SCRIPT_PATH=${0%/*}
 cd ${SCRIPT_PATH}
+
+# Abort immediately if the database is not SQLite
+if [ ! -e db.sqlite3 ]
+then
+    echo "Aborting with no changes, because could not detect db.sqlite3."
+    exit
+fi
 
 # Remove the existing database and migration files if there are any.
 rm db.sqlite3
@@ -15,5 +25,5 @@ python manage.py migrate
 # Create an admin user called admin with password fibble.
 echo "from django.contrib.auth.models import User; User.objects.create_superuser('admin', 'admin@example.com', 'fibble')" | python manage.py shell
 
-# Populate the database with the starter content - i.e reference experiement.
-python manage.py reset_db
+# Further populate the database with the reference experiment.
+python manage.py pop_with_ref_exp
